fixed crashes
