+++ app/src/main/java/colosseum19/a300dpi/colosseum2k19/Adapters/FixtureAdapter.java
@@ -80,7 +80,6 @@ public class FixtureAdapter extends RecyclerView.Adapter<FixtureAdapter.FixtureH
     public void setFixtureData(ArrayList<Fixture> data, boolean isEmpty) {
         progressDialog.cancel();
         if(!isEmpty){
-            Log.d("DATA TEST", data.toString());
             this.fixtureList = data;
             notifyDataSetChanged();
         }else{
