+++ app/src/main/java/colosseum19/a300dpi/colosseum2k19/HomeActivity.java
@@ -1,6 +1,8 @@
 package colosseum19.a300dpi.colosseum2k19;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
@@ -20,13 +22,15 @@ import com.google.firebase.messaging.FirebaseMessaging;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import colosseum19.a300dpi.colosseum2k19.API.BackupApi;
 import colosseum19.a300dpi.colosseum2k19.Fragments.AboutFragment;
 import colosseum19.a300dpi.colosseum2k19.Fragments.EventsFragment;
 import colosseum19.a300dpi.colosseum2k19.Fragments.FixturesFragment;
 import colosseum19.a300dpi.colosseum2k19.Fragments.ScoresFragment;
+import colosseum19.a300dpi.colosseum2k19.Interfaces.DateInterface;
 import colosseum19.a300dpi.colosseum2k19.Utilities.ConstantsStorage;
 
-public class HomeActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener, BottomNavigationView.OnNavigationItemReselectedListener {
+public class HomeActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener, BottomNavigationView.OnNavigationItemReselectedListener, DateInterface {
 
     //Keys for Fragments
     private static String EVENTS_FRAG = "events_frag";
@@ -48,6 +52,7 @@ public class HomeActivity extends AppCompatActivity implements BottomNavigationV
     private FixturesFragment fixturesFragment;
     private ScoresFragment scoresFragment;
     private AboutFragment aboutFragment;
+    private BackupApi backupApi;
 
     private String currentlySelected = EVENTS_FRAG;
 
@@ -64,6 +69,9 @@ public class HomeActivity extends AppCompatActivity implements BottomNavigationV
         setContentView(R.layout.activity_home);
         ButterKnife.bind(this);
         setSupportActionBar(toolbar);
+
+        backupApi = new BackupApi();
+        backupApi.getDate(getApplicationContext(), this);
 
         fragmentManager = getSupportFragmentManager();
         bottomNavigationView.setOnNavigationItemSelectedListener(this);
@@ -135,6 +143,18 @@ public class HomeActivity extends AppCompatActivity implements BottomNavigationV
             }
         }
 
+    }
+
+    @Override
+    public void setDate(String date, boolean isSuccess) {
+        if(isSuccess){
+            SharedPreferences sharedPreferences = getSharedPreferences(getString(R.string.shared_pref_name), Context.MODE_PRIVATE);
+            SharedPreferences.Editor editor = sharedPreferences.edit();
+            editor.putString(getString(R.string.day_key), date);
+            editor.apply();
+        }else{
+            Toast.makeText(this, "Failed to get current date", Toast.LENGTH_SHORT).show();
+        }
     }
 
     @Override
